+++ changelog.md
@@ -1,9 +1,15 @@
 # Changelog
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+Types of changes: Added, Changed, Deprecated, Removed, Fixed, Security
+
 ## [Unreleased]
+
+### Changed
+
+- The method 'db.AMPODatabase.get_db' is not longer a 'classmethod'
 
 ## [0.2.2] - 2024-02-17
 
@@ -15,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- The method 'update_expiration_value' was added for worker.
+- The 'update_expiration_value' method has been added for worker.
 - Was changed behavior for TTL index.